requirements.txt
ultralytics==8.3.144
requests==2.31.0
mysql-connector-python==9.3.0
